reorganize google tag manager, try removing revalidate option
