+++ frontend/lib/constants.ts
@@ -1,3 +1,5 @@
+export const GOOGLE_TAG_MANAGER_ID = "G-WW9E92XGYR";
+
 export const PUBLIC_CALENDAR_URL =
   "https://outlook.office365.com/owa/calendar/e17014b3d3194bd28aeac7f6981e7637@smgschool.org/968a59748b1341389f01860bcc52b3694203663748153909325/calendar.html";
